refactor: thermo-nuclear cleanup round 2
- manager.ts: collapse the triple internal-plugin emit guard into a
  single record() helper — one rule, one comment, one place to change
- defaultPluginPackages.ts: drop the silent import.meta.url resolution
  fallback; a host that forgot to declare a plugin dep should fail
  loudly at boot, not resolve through accidental hoisting
- cli.ts: extract both app compositions (folder + workspaces mode) and
  their plugin-runtime plumbing into modeApps.ts. cli.ts shrinks from
  1012 to 317 lines and is again just CLI entry/arg/static concerns;
  public symbols re-exported for existing importers
