+++ packages/cli/tsup.config.ts
@@ -5,6 +5,7 @@ export default defineConfig({
     "src/index.ts",
     "src/server/cli.ts",
     "src/server/localWorkspaces.ts",
+    "src/server/modeApps.ts",
     "src/server/pluginDiscovery.ts",
     "src/server/pluginFrontRuntime.ts",
   ],
